stream align: Add cross stream sync tag share
Now it is possible to process multiple receive streams in parallel. In
case on stream misses a tag, the other streams share this tag in order
to improve reliability and ensure tag alignment.

This addition revealed a severe bug. xcorr_tagger tagged past samples
that where already consumed in an earlier call to work. Since these tags
are usually only off by 1 sample, the corresponding tag is moved to the
appropriate position, the first position.
In essence this fix results in a rewrite of this block.